Drop Python 3.4
At this point Python 3.4 is basically at its EOL. So go ahead and drop
it. Though the code here will likely work fine there. We have no
interest in testing or maintaining that platform. Further we'd like to
start testing on Python 3.7. This will free up some CI resources for
that effort.