Exclude branch builds for non-`master` branches
When building a PR off of a feature branch that we created, the CIs
trigger a branch build and a PR build. These are largely redundant and
the former does not test the merge commit. This should drop the branch
build, but keep the PR build intact. Further this should continue to
build pushes to the `master` branch.