Perform the extension build in-place
As other issues filed about Cython's coverage support mention that it
looks for C files alongside the Cython files, presumably ensuring the C
files show up next to the Cython files should help. In an attempt to do
this, try building the extension modules in-place. This should ensure
that the C files show up next to their Cython counterparts.