Create `cybuffer` extension type
Implements the `cybuffer` extension type. This supports Python's old
buffer protocol (Python 2 only) and the (new) buffer protocol (Python
2+). The value being that this can be comfortably passed in anywhere
either buffer protocol is needed. As some Python 2 functions (e.g.
compressors, hashes, etc.) only support the old buffer protocol, this
should work nicely with them. In other cases a Python 2 function may
well with either protocols or have a slight preference for one or the
other. As this supports both buffer protocols, it will coerce easily to
whichever one works best for the algorithm. In Python 3 only the (new)
buffer protocol is supported, which this works fine with as well.

This works by trying to coerce the input data provided into the (new)
buffer protocol falling back to the old one when nothing else works. In
either case, it takes a `memoryview` of the resulting coercion. This
standardizes the view of the data between the two interfaces. Also
provides some information that is needed for supporting the (new) buffer
protocol. Of course this information can easily be repurposed to support
the old buffer protocol. As the `memoryview` manages all underlying
allocated buffers, there is no need to do any special handling when our
object scopes out as the `memoryview` will clean that up on the way out.

Generally when exposing the data to the (new) buffer protocol, we just
copy over the `Py_buffer` object held by the `memoryview` for the
consumer. Though we do replace `obj` to point to our `self` instead to
ensure our object doesn't scope out. Otherwise our object and the
`memoryview` could get deallocated resulting in segfaults and other
unpleasant outcomes. However sometimes we have to do a little more to
make the buffer more user friendly. For instance if a builtin `array`
object is provided as data, we have to handle some peculiarities of it.

In particular the builtin `array` supports a `unicode` type (i.e.
`Py_UNICODE`), which has been a bit broken and doesn't necessarily work
well with other (new) buffer protocol consumers (e.g. NumPy).
Specifically the `unicode` type may be `Py_UCS4` or `Py_UCS2` depending
on how Python was built. The former is supported by NumPy, but the
latter is not. While this type has been deprecated by Python, it still
exists in Python 2 and 3 (though will likely be pulled out in Python 4).
So we try to support it as best we can by using the same strategy
`memoryview`s already use with `bytes`. Namely casting it to an unsigned
integer type. Thus we cast `Py_UCS2` to `uint16_t` and `Py_UCS4` to
`uint32_t`. These can easily be handled by NumPy or any other consumer
of `memoryview`s. Also its consistent with `cybuffer` being a data
container as opposed to some way to manipulate text encodings.

On Python 2 we have the added problem that builtin `array`s do not
support the (new) buffer protocol. However creating the `memoryview`
through the old buffer protocol results in a loss of type information as
it represents everything as unsigned bytes. To fix this, we patch the
copied buffer to include information about the type of `array` we were
provided. Specifically this means we update the `format` string,
`itemsize`, `shape`, and `strides` set in the `Py_buffer` for the
consumer. This patching is relatively fast as most of the operations
happen at the C level with very minimal usage of Python objects or the
Python API. As a result, `array` types can be passed into `cybuffer` and
be used on Python 2/3 as if they supported the (new) buffer protocol on
both platforms.

As all information is available from the `memoryview`, there really is
no issue filling out that information into the user provided buffer.
It's just a matter of making sure we don't accidentally write out
components that were not requested. This would be particularly bad for
some pointers that might not be allocated. So we check the flags to see
which pieces should be written out and only write those parts. The rest
are set to NULL. Note this has no effect on the non-pointer types, which
we write out anyways. All of those are required regardless anyways. Also
we have no internal content that needs to be kept around. So that field
is set to NULL as well.

When it comes to the other requirements like in what ways the data needs
to be contiguous and whether it is writable, we have no real control
over that as we are not maintaining the buffer ourselves. So instead we
simply check to see if the data we have meets the user's requests. If it
does, then we proceed to filling out the `Py_buffer` object. However if
any of these doesn't match their expectations, we raise a `BufferError`
as the (new) buffer protocol specifies. This may trigger copies
downstream depending on what exactly was requested in how downstream
buffers want to handle it.

Exposing the old buffer protocol involves using the legacy Cython dunder
methods. These pretty much just expose the `memoryview`s buffer to the
consumer after some basic checks. All of them require that the array is
contiguous. If that isn't upheld, they raise. While it may be possible
to handle some of the more advanced slicing cases using multiple
segments, this becomes rather complex and raises various questions that
need to be answered.

For what it is worth NumPy seems to agree with this sentiment as they
did not implement them either. More accurately they have a comment that
states it was removed, which appears to predate SciPy! In fact it is
likely the weakness of the old buffer interface in this regard that gave
way to the new.

In order to not only provide a reasonably transportable container of
data for the different buffer protocols, but also to provide some basic
level of functionality, expose some of the internal state as `readonly`
attributes for others to inspect the metadata. Where attributes do not
suffice, provide small methods exposed as properties to coerce C types
into things familiar to Python programmers (e.g. C pointer of shape to a
Python tuple). To allow typical methods to work more nicely, provide a
few dunder methods for use with Python to access things like length.
These all line up with their names on the `memoryview` interface of
Python 2 and Python 3. Though should expose a few more than what Python
2's `memoryview` has out-of-the-box.

As some cases need another simple Python native representation of the
binary data, implement the `tobytes` method. As our type patching does
not effect the underlying buffer, we simply rely on our `memoryview` of
the data type unpatched so as to get access to its `Py_buffer` for use
with the (new) buffer protocol. Then we allocate a `bytes` object of the
right size that is otherwise empty and get access to its underlying C
representation as a `char*` so that we can fill it. Using
`PyBuffer_ToContiguous`, we perform a direct copy from the
`memoryview`'s buffer into the `bytes` object. Returning the filled
`bytes` object to the user. The buffer protocol is able to handle any
striding we throw at it without generating intermediates making this
very fast.  As the whole thing uses the C API, there is very little
overhead from CPython other than the usual reference counting. This ends
up happening at a comparable speed to NumPy's own `tobytes`, which may
in fact use the same mechanism.