Use Cython `memoryview` to implement get/set item
Adds a little bit of code to leverage Cython's `memoryview` objects to
enable indexing the `cybuffer` object to get a new one on a restricted
view or to set values in it. In both cases, leverage the buffer protocol
to get a Cython `memoryview` onto our data. Thus it applies our castings
as well. Cython's `memoryview` is pretty capable as it allows us to make
subselections onto the data. In get item, construct a new `cybuffer`
around Cython's `memoryview` when returning these subselections. Thus
users just see new `cybuffer`s and the Cython `memoryview` is an
implementation detail. In set item, just create a Cython `memoryview`
of our data and use their set item to assign the value. Should handle
the casting we apply nicely.