Implement tolist
Adds an implementation of `tolist` for `cybuffer`. Uses a recursive
helper function in basically straight C to build lists of values, which
are then reinserted into other enclosing lists. Makes use of Cython's
`memoryview` instance to handle the get item calls as we recurse down.
Preallocates all lists with `NULL` values and sets them with actual
objects using Python's C API to make the list construction faster.
Requires use of increment reference calls to handle reference stealing
in the context of Cython's automatic reference counting. This is slower
than NumPy's implementation. That said, building large Python lists of
data is probably the wrong way to go. Should be fine for getting a rough
idea of what's in the memory shared with `cybuffer` without resorting to
other libraries.