Skip memory deallocation if not Python 2
This memory is only allocated on Python 2 in the special case where a
builtin `array` is used and its type is not unsigned char or mappable to
it (e.g. `"c"`). As such, we can avoid these checks entirely if we are
not running on Python 2. So add a check for Python 2. This will ensure
this code block is dropped on Python 3.