Add `hex` method to `cybuffer`
Provides a Cython/Python method to get the hex representation of the
data in a string (bytes or unicode on Python 2 or 3 respectively) much
like the method of the Python 3 `memoryview`.  Performance is very close
to similar strategies with a `bytes` like object or `memoryview`
respectively on either Python version using equivalent strategies. Also
test this method against Python 3's `memoryview`'s `hex` method.
Fallback to `binascii.hexlify` on Python 2.