Optimize `cybuffer`'s method `tolist`
Makes some improvements to `tolist` by rewriting the underlying utility
function to operate directly on the raw pointer to the data instead of
using Cython's `memoryview` object. Overall improves the speed of
`tolist` by 4-fold. Also removes reliance on Cython's `memoryview` type
for the `tolist` method, which saves a fair bit of overhead.