Support `suboffsets` in `tolist`
Update the `tolist` implementation to handle `suboffsets`. If any values
in `suboffsets` are negative, skip applying them. Further if
`suboffsets` is `NULL`, treat it as if all `suboffsets` are negative.
Thus the pre-`suboffsets` implementation's behavior is preserved when
the `suboffsets` are `NULL` as is the case with our current test cases.