Cast exotic formats to unsigned integers
If `Struct` is unable to recognize the format, we are going to be unable
to create `memoryview`s or do other things with the data easily. So cast
these to unsigned integers of the right size. This admittedly does not
handle all cases, but this will handle things like datetime/timedelta
data. For all cases, this cannot handle, raise an exception.