Check (new) buffer protocol support on Python 2
Skip checking to see if the object supports the (new) buffer protocol
unless we are on Python 2. On Python 2/3, `PyObject_GetBuffer` already
raises a `TypeError` with an informative message. This is also true of
`PyBuffer_FromObject` on Python 2. So it is better to allow Python's
`TypeError` messages to propagate up instead of adding our own. This
keeps the code simple for our use case with the same end effect. Also
makes the Python 3 case a little cleaner.

Of course on Python 2 if the object doesn't support (new) buffer
protocol, it's good to figure that upfront as the check is cheap
compared to catching an exception. That way we can fallback to the
buffer protocol simply. Instead of carrying around a new object for
storing the result, just overwrite `data` as we already have a copy of
it earlier and this won't affect anything outside of this function. Any
exceptions that would be propagated on the Python 2 side are handled or
not just as before.