Skip creating a `memoryview` from `obj`
At this stage, there is no real value to creating a `memoryview` from
our data to create an object vs. just using the buffer protocol to fill
out our buffer. Under the hood, the `PyMemoryView_FromObject` call is
doing exactly what we are doing and providing little more. Filling out
the buffer directly avoids creating an intermediate object between our
buffer and the data. Not to mention this should make it easier for us to
get access to the underlying object used to create the buffer. Also
should cutdown on the overhead of generating a `cybuffer` object.