Only patch-up `array` buffer on Python 2/3
When initializing `cybuffer` with an `array, ensure that we only patch
up the `format` and associated attributes for the `array` buffer on
Python 2/3. This workaround is needed for Python 2 as the `array` type
does not support the (new) buffer protocol. So we need to use the old
buffer protocol and then the (new) buffer protocol with that result.
This loses information about the `array` in the process, which is the
bulk of the reason this is needed. Python 3's `array` supports the (new)
buffer protocol; so, we can largely ignore these workarounds. However on
Python 2/3, we have the `"u"` (unicode) format for the `array`, which is
a little broken. So we workaround those issues by casting the array to
an unsigned integer of equivalent width to the unicode type. However the
`"u"` format for `array` is deprecated and scheduled for removal in
Python 4. So there should be no need to do any of this work on Python 4.
As such, check whether Python 2/3 is in use when checking if the data is
provided is an `array`. That way this whole branch will automatically be
removed on Python 4 when it lands.