Define `Py_UNICODE_SIZE` at Cython compile-time
Previously we dealt with `Py_UNICODE_SIZE` as a C compile-time constant,
which would inject the value during the C preprocessor pass. Later
compilation optimizations by the C compiler likely removed unneeded
branches caused by this. Though this is pretty needless as Cython is
generating these branches that we later drop. Not to mention Cython
sometimes assigns constants like these to intermediate or stack local
variables, which can hinder compiler optimizations. It would be better
yet to have Cython not define these branches at all.

To that end, we make `Py_UNICODE_SIZE` a Cython compile-time constant
that is set in `config.pxi` along with other compile-time constants. As
a result, Cython injects these directly in `cybuffer.pxd` and
`cybuffer.pxi` before even performing any compilation. Once it goes to
compile the Cython code to C, it recognizes `Py_UNICODE_SIZE` is a
compile-time constant and evaluates all expressions it is used in. As a
result, the Cython branches with `Py_UNICODE_SIZE` are eliminated during
compile-time leaving only one unconditional code path in C.

In order to discern what `Py_UNICODE_SIZE` should be at the Python
level, we inspect the `itemsize` of an empty unicode (i.e. `"u"`)
`array`'s `itemsize` in `setup.py`. Based on this we can determine
`Py_UNICODE_SIZE`'s value. We then inject this into `config.pxi`. This
value is only inspected and added for Python 2/3 as the unicode `array`
is scheduled for removal in Python 4. As we only check this value
within a branch of code that should be removed on Python 4 anyways,
leaving `Py_UNICODE_SIZE` undefined shouldn't be an issue for Python 4
builds.

While `Py_UNICODE_SIZE` is defined in Python headers, we don't need to
be worried about our definition of `Py_UNICODE_SIZE` clashing with it.
The reason being Cython guarantees the assigned value will injected into
the source. So the definition will have already been eliminated by the
time the C code is generated. Further we only use `Py_UNICODE_SIZE` in
comparisons to other constants. So Cython will inject the value and then
evaluate these comparisons resulting in them being turned into always
`True` or `False` branches. These in turn are optimized away.