Drop character array format assignment on Python 2
On Python 2, the `format` of all `array`s is `uint8` (i.e. `"B"`). The
reason being that all `array`s on Python 2 only support the old buffer
protocol. As such they need to be converted through that path to get
access to use them with the (new) buffer protocol. Doing so results in
such objects being treated as if they are of `uint8` type. Since that is
how we want character (i.e. `"c"`) arrays treated on Python 2 (they
don't exist on Python 3), then there is nothing we need to here. This
includes changing the `_format`. So drop that line.