Check itemsize before adjusting array on Python 2
On Python 2 when handling the `array` case, the only reason to worry
about adjusting the shape and strides is if the `itemsize` changed as a
consequence of casting back to the original type. If the itemsize is the
same, which on Python 2 is 1 as a consequence of needing to use the old
buffer protocol for `array`s, then there is no need to perform this
casting. Hence we update the Python 2 shape and stride adjustment case
to check to see if the `itemsize` is different from `1`, which is now
possible as we assign the `itemsize` when handling the casting above.
This is fast check since it is comparing to C integers and only matters
for Python 2 as the code is removed for newer Python versions.