Simplify `array`'s shape and strides adjustment
When adjusting the `shape` and `strides` to take into account Python 2's
`array`'s recasting of type, simplify the code a bit knowing the Python
2 `array` must be 1-D. As such there is no need to find the last
dimension. Instead just update the `shape` and `strides` at index `0` as
that is the first and last dimension. This simplifies the code a bit,
improves C code generated from it, and clarifies the code's intent.