Create the `getbuffer` function
Adds the `getbuffer` function for Python 2. This maps nicely to the
builtin `buffer` function and NumPy's `getbuffer` function. Basically
uses the old buffer protocol to try and return a read-write buffer
falling back to a read-only buffer if that fails. If both fail, all
exceptions are propagated to the end user.