Test `getbuffer` on Python 2
Provides a test with `bytes` and `bytearray` on Python 2 using
`getbuffer`. Makes sure that we get a `buffer` instance in all cases.
Also test that the content matches the data provided. Finally make sure
that a `memoryview` of the `buffer` instance has the same permissions as
a `memoryview` onto the data (i.e. write permissions are preserved).