Use `PyObject_AsWriteBuffer` to check for writing
Apparently `PyBuffer_FromReadWriteObject` does not raise exceptions in
all the cases that it should. Namely it will happily provide a
read-write `buffer` for a `bytes` object, which should not be possible.
To fix this, use `PyObject_AsWriteBuffer` to check to see if the buffer
is writable as this function seems to raise exceptions correctly. As
Cython does not include `PyObject_AsWriteBuffer`, add our own `extern`
for it. If `PyObject_AsWriteBuffer` raises a `TypeError`, catch it and
use `PyBuffer_FromObject` to provide a read-only `buffer`. However if no
exception is raised, use `PyBuffer_FromReadWriteObject` to create a
read-write `buffer`. This seems to work correctly based on our testing.

ref: https://docs.python.org/2/c-api/objbuffer.html