Perform string comparisons directly for array case
When handling the `array` case, go ahead and copy the `itemsize` and
`format` over to start with on Python 2. This way Python 2/3 cases are
handled basically the same after this point. Then use `_format` to
perform string comparisons in C instead of Python. As the strings are
all short, these are very fast. This avoids the overhead of Python
string comparison.

Update the unicode strategy to handle the fact that UCS2 is represented
by the `"u"` format and UCS4 is represented by the `"w"` format. Specify
the type casting for each case using the appropriate unsigned integer
width type as before. As a result the format check now occurs at
run-time instead of build-time, this should make binaries more portable.

Explicitly check for the character array format `"c"` on Python 2. In
this case patch the format, to be the unsigned char format `"B"`. This
is basically what we did before. However now we do so explicitly. As
this type only exists on Python 2 (not Python 3), there is no need to
handle this case for other Python version. So include `PY2K` in the
conditional to ensure this check is not included on Python 3.