Handle different unicode typecodes
In some cases the unicode typecode `"w"` shows up, but it is not well
documented and is on its way out as well. That said, the code assumed
this to be UCS4. This may very well be true, but it is a bit unclear.
Similarly on Python 2, it seems that the unicode typecode `"u"` can mean
either UCS2 or UCS4. Given the confusion about the unicode typecodes,
lump all comparisons of them together and simply handle casting based on
`itemsize`. This should avoid getting into any technical issues of these
legacy types while still maintaining the intended behavior (casting to
an appropriately sized unsigned integer). While we are at it, go ahead
skip defining these legacy types as they don't have as a clear of a
mapping as one might initially think.