Rewrite `array` type patching with `switch`/`case`
Instead of using `strcmp`, simply access the first `char` of the
`format` string and assign it to `fmt` a `char` type. Then compare the
different values that `fmt` can take on. Cython turns this into a
`switch`/`case` block with relatively little interference. So the
compiler can easily optimize this into the appropriate form.